+++ src/autocapture-utils.ts
@@ -4,7 +4,9 @@
  * @returns {string} the element's class
  */
 import { AutocaptureConfig } from 'types'
-import { _each, _includes, _isNull, _isString, _isUndefined, _trim, logger } from './utils'
+import { _each, _includes, _trim, logger } from './utils'
+
+import { _isNull, _isString, _isUndefined } from './type-utils'
 
 export function getClassName(el: Element): string {
     switch (typeof el.className) {
