+++ src/extensions/exception-autocapture/error-conversion.ts
@@ -9,7 +9,8 @@ import {
     isPrimitive,
 } from './type-checking'
 import { defaultStackParser, StackFrame } from './stack-trace'
-import { _isNumber, _isString, _isUndefined } from '../../utils'
+
+import { _isNumber, _isString, _isUndefined } from '../../type-utils'
 
 /**
  * based on the very wonderful MIT licensed Sentry SDK
