+++ src/extensions/exception-autocapture/index.ts
@@ -1,8 +1,10 @@
-import { _isArray, _isObject, _isUndefined, logger, window } from '../../utils'
+import { logger, window } from '../../utils'
 import { PostHog } from '../../posthog-core'
 import { DecideResponse, Properties } from '../../types'
 import { ErrorEventArgs, ErrorProperties, errorToProperties, unhandledRejectionToProperties } from './error-conversion'
 import { isPrimitive } from './type-checking'
+
+import { _isArray, _isObject, _isUndefined } from '../../type-utils'
 
 const EXCEPTION_INGESTION_ENDPOINT = '/e/'
