+++ src/extensions/exception-autocapture/type-checking.ts
@@ -1,4 +1,4 @@
-import { _isFunction, _isNull, _isObject, _isUndefined } from '../../utils'
+import { _isFunction, _isNull, _isObject, _isUndefined } from '../../type-utils'
 
 export function isEvent(candidate: unknown): candidate is Event {
     return !_isUndefined(Event) && isInstanceOf(candidate, Event)
