+++ src/loader-exception-autocapture.ts
@@ -1,5 +1,6 @@
 import { extendPostHog } from './extensions/exception-autocapture'
-import { _isUndefined } from './utils'
+
+import { _isUndefined } from './type-utils'
 
 const win: Window & typeof globalThis = _isUndefined(window) ? ({} as typeof window) : window
