+++ src/loader-recorder-v2.ts
@@ -7,7 +7,8 @@ import rrwebRecord from 'rrweb/es/rrweb/packages/rrweb/src/record'
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 import { getRecordConsolePlugin } from 'rrweb/es/rrweb/packages/rrweb/src/plugins/console/record'
-import { _isUndefined } from './utils'
+
+import { _isUndefined } from './type-utils'
 
 const win: Window & typeof globalThis = _isUndefined(window) ? ({} as typeof window) : window
