+++ src/loader-recorder.ts
@@ -7,7 +7,8 @@ import rrwebRecord from 'rrweb-v1/es/rrweb/packages/rrweb/src/record'
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 import { getRecordConsolePlugin } from 'rrweb-v1/es/rrweb/packages/rrweb/src/plugins/console/record'
-import { _isUndefined } from './utils'
+
+import { _isUndefined } from './type-utils'
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
